requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
decouple==0.0.7
distro==1.9.0
Django==5.0.5
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.25.2
pdf2image==1.17.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
